update other queries and pages
