+++ details/main.py
@@ -8,7 +8,7 @@ from details.isotherms import plot_isotherms, get_widom_df
 from details.dft_info import plot_energy_steps
 from details.structure import structure_jsmol
 from details.utils import get_mat_id, get_details_title, get_geom_table, get_appl_table, get_title
-from pipeline_config import get_mat_nodes_dict
+from pipeline_config import get_mat_dict
 
 load_profile()
 
@@ -19,23 +19,22 @@ class DetailView():
 
     def __init__(self):
         self.mat_id = get_mat_id()
-        self.mat_nodes_dict = get_mat_nodes_dict(self.mat_id)
-        print(">> Display details of MAT_ID:", self.mat_id, self.mat_nodes_dict['orig_cif'])
+        self.mat_dict = get_mat_dict(self.mat_id)
+        print(">> Display details of MAT_ID:", self.mat_id)
 
     def title_col(self):
         col = pn.Column(width=700)
-        col.append(pn.pane.Markdown(get_details_title(self.mat_nodes_dict['orig_cif'])))
+        col.append(pn.pane.Markdown(get_details_title(self.mat_dict)))
         return col
 
     def structure_col(self):
-        nodes = self.mat_nodes_dict
         col = pn.Column(sizing_mode='stretch_width')
-        col.append(get_title('Cell optimized structure', uuid=nodes['opt_cif_ddec'].uuid))
-        col.append(pn.pane.Bokeh(structure_jsmol(nodes['opt_cif_ddec'])))
-        col.append(get_title('Energy profile during cell optimization', uuid=nodes['dftopt'].uuid))
-        col.append(pn.pane.Bokeh(plot_energy_steps(dftopt_out=nodes['dftopt'])))
-        col.append(get_title('Geometric properties', uuid=nodes["opt_zeopp"].uuid))
-        col.append(pn.pane.Markdown(get_geom_table(nodes["opt_zeopp"])))
+        col.append(get_title('Cell optimized structure', uuid=self.mat_dict['opt_cif_ddec'].uuid))
+        col.append(pn.pane.Bokeh(structure_jsmol(self.mat_dict['opt_cif_ddec'])))
+        col.append(get_title('Energy profile during cell optimization', uuid=self.mat_dict['dftopt'].uuid))
+        col.append(pn.pane.Bokeh(plot_energy_steps(dftopt_out=self.mat_dict['dftopt'])))
+        col.append(get_title('Geometric properties', uuid=self.mat_dict["opt_zeopp"].uuid))
+        col.append(pn.pane.Markdown(get_geom_table(self.mat_dict["opt_zeopp"])))
         return col
 
     def properties_col(self):
@@ -44,16 +43,16 @@ class DetailView():
         col.append(pn.pane.Bokeh(plot_isotherms(self.mat_id), sizing_mode='stretch_width'))
         col.append(pn.pane.Markdown("## All computed Henry's coefficients (mol/kg/Pa)"))
         with pd.option_context('display.max_colwidth', None):
-            col.append(get_widom_df(self.mat_nodes_dict, select="kh").to_html(escape=False))
+            col.append(get_widom_df(self.mat_dict, select="kh").to_html(escape=False))
         col.append(pn.pane.Markdown("## All computed Heat of adsorption @ infinite dilution (kJ/mol)"))
         with pd.option_context('display.max_colwidth', None):
-            col.append(get_widom_df(self.mat_nodes_dict, select="hoa").to_html(escape=False))
+            col.append(get_widom_df(self.mat_dict, select="hoa").to_html(escape=False))
         return col
 
     def applications_col(self):
         col = pn.Column(sizing_mode='stretch_width')
         col.append(pn.pane.Markdown('## Numerical values for all the applications'))
-        col.append(pn.pane.HTML(get_appl_table(self.mat_nodes_dict)))
+        col.append(pn.pane.HTML(get_appl_table(self.mat_dict)))
         return col
