+++ pipeline_config/__init__.py
@@ -1,18 +1,61 @@
 """Defining my color palette for visualization."""
-import collections
-import yaml
 import re
 import os
+import collections
+import yaml
+import pandas as pd
 from os.path import join, dirname, realpath
 from frozendict import frozendict
 from functools import lru_cache
 from aiida.orm.querybuilder import QueryBuilder
 from aiida.orm import Node, Dict, Group, WorkChainNode
 
+TAG_KEY = "tag4"
+GROUP_DIR = "discover_curated_cofs/"
 CONFIG_DIR = join(dirname(realpath(__file__)), "static")
 EXPLORE_URL = os.getenv('EXPLORE_URL', "https://dev-www.materialscloud.org/explore/curated-cofs")
+AIIDA_LOGO_URL = "select-figure/static/images/aiida-128.png"
+CO2_LOGO_URL = 'select-figure/static/images/co2-128.png'
 
 TAG_KEY = 'tag4'
+
+
+# Load and update profile for dockers
+def update_config():
+    """Add AiiDA profile from environment variables, if specified"""
+    from aiida.manage.configuration import load_config
+    from aiida.manage.configuration.profile import Profile
+
+    profile_name = os.getenv("AIIDA_PROFILE")
+    config = load_config(create=True)
+    if profile_name and profile_name not in config.profile_names:
+        profile = Profile(
+            profile_name, {
+                "database_hostname": os.getenv("AIIDADB_HOST"),
+                "database_port": os.getenv("AIIDADB_PORT"),
+                "database_engine": os.getenv("AIIDADB_ENGINE"),
+                "database_name": os.getenv("AIIDADB_NAME"),
+                "database_username": os.getenv("AIIDADB_USER"),
+                "database_password": os.getenv("AIIDADB_PASS"),
+                "database_backend": os.getenv("AIIDADB_BACKEND"),
+                "default_user": os.getenv("default_user_email"),
+                "repository_uri": "file://{}/.aiida/repository/{}".format(os.getenv("AIIDA_PATH"), profile_name),
+            })
+        config.add_profile(profile)
+        config.set_default_profile(profile_name)
+        config.store()
+
+    return config
+
+
+def load_profile():
+    import aiida
+
+    update_config()
+    aiida.load_profile()
+
+
+load_profile()
 
 # Get quantities and applications
 with open(join(CONFIG_DIR, "quantities.yml"), 'r') as f:
@@ -66,44 +109,52 @@ gasses = collections.OrderedDict([(q['label'], q) for q in gasses_list])
 
 
 # Get queries
-@lru_cache(maxsize=128)
+#@lru_cache(maxsize=128)
 def get_data_aiida(quantitites):
-    """Query the AiiDA database
-
-    :param quantities: tuple of quantities to project
-
-    TODO: the group version needs to be better rationalized!
-    """
+    """Query the AiiDA database for a list of quantities (other appl results)."""
 
     qb = QueryBuilder()
-    qb.append(Group, filters={'label': {'like': r'curated-cof\_%\_v_'}}, tag='curated_groups')
-    qb.append(Node,
-              project=['label', 'extras.name_conventional', 'extras.class_material'],
-              filters={'extras.{}'.format(TAG_KEY): 'orig_cif'},
-              with_group='curated_groups')
+    qb.append(Group,
+              filters={'label': {
+                  'like': GROUP_DIR + "%"
+              }},
+              tag='g',
+              project=['extras.mat_id', 'extras.name_conventional', 'extras.class_material'])
 
     for q in quantitites:
         qb.append(Dict,
                   project=['attributes.{}'.format(q['key'])],
                   filters={'extras.{}'.format(TAG_KEY): q['dict']},
-                  with_group='curated_groups')
+                  with_group='g')
 
     return qb.all()
 
 
-def get_mat_nodes_dict(mat_id):
-    """Given a MAT_ID return a dictionary with all the tagged nodes for that material."""
+def get_mat_dict(mat_id):
+    """Given a curated-cof label, queries the group and returns a dictionary with tags as keys and nodes as values.
+    If multiple version are available qb.all()[0][0] shuld take the last one computed.
+    """
 
+    mat_dict = {}
+
+    # get all nodes
     qb = QueryBuilder()
-    qb.append(Group, filters={'label': {'like': r'curated-___\_{}\_v_'.format(mat_id)}}, tag='curated_groups')
-    qb.append(Node, filters={'extras': {'has_key': TAG_KEY}}, with_group='curated_groups')
+    qb.append(Group, filters={'label': {'like': GROUP_DIR + mat_id}}, tag='group')
+    qb.append(Node, project=['extras.{}'.format(TAG_KEY), '*'], with_group='group')
 
-    mat_nodes_dict = {}
-    for q in qb.all():
-        n = q[-1]  # if more groups are present with different versions, take the last: QB sorts groups by label
-        mat_nodes_dict[n.extras[TAG_KEY]] = n
+    for k, v in qb.all():
+        mat_dict[k] = v
 
-    return mat_nodes_dict
+    # get extra info
+    qb = QueryBuilder()
+    qb.append(Group, filters={'label': {'like': GROUP_DIR + mat_id}}, project=['*'])
+    g = qb.all()[0][0]
+    mat_dict['mat_id'] = g.extras['mat_id']
+    mat_dict['name_conventional'] = g.extras['name_conventional']
+    mat_dict['doi_ref'] = g.extras['doi_ref']
+    mat_dict['workflow_version'] = g.extras['workflow_version']
+
+    return mat_dict
 
 
 @lru_cache(maxsize=8)
@@ -156,8 +207,93 @@ def get_isotherm_nodes(mat_id):
     return gas_dict
 
 
-# Get color palette
+def provenance_link(uuid, label=None):
+    """Return representation of provenance link."""
 
+    if label is None:
+        label = "Browse provenance\n" + str(uuid)
+
+    return "<a href='{url}/details/{uuid}' target='_blank'><img class='provenance-logo' src={logo_url} title='{label}'></a>".format(  # noqa
+        url=EXPLORE_URL, uuid=str(uuid), label=label, logo_url=AIIDA_LOGO_URL)
+
+
+def detail_link(mat_id):
+    """Return representation of workflow link."""
+    return "<a href='detail?mat_id={}' target='_blank'><img class='provenance-logo' src='{}'></a>".format(
+        mat_id, CO2_LOGO_URL)
+
+
+def doi_link(mat_dict):
+    """Return the DOI link of the article."""
+    name = mat_dict['name_conventional']
+    doi = mat_dict['doi_ref']
+    return "<a href='https://doi.org/{}' target='_blank'>{}</a>".format(doi, name)
+
+
+@lru_cache()
+def get_db_nodes_dict():
+    """Given return a dictionary with all the curated materials having the material label as key, and a dict of
+    curated nodes as value.
+    IMPROVED FOR SPEED!
+    """
+
+    qb = QueryBuilder()
+    qb.append(Group, filters={'label': {'like': GROUP_DIR + "%"}}, tag='g', project=['extras'])
+    qb.append(Node, filters={'extras': {'has_key': TAG_KEY}}, with_group='g', project=[f'extras.{TAG_KEY}', 'uuid'])
+
+    db_nodes_dict = {}
+    for q in qb.all():  # q = [group-label, group-extras, node-tag, node-uuid]
+        mat_id = q[0]['mat_id']
+        if mat_id not in db_nodes_dict:
+            db_nodes_dict[mat_id] = {
+                'name_conventional': q[0]['name_conventional'],
+                'doi_ref': q[0]['doi_ref'],
+                'workflow_version': q[0]['workflow_version'],
+            }
+        db_nodes_dict[mat_id][q[1]] = q[2]
+
+    return db_nodes_dict
+
+
+@lru_cache()
+def get_table():
+    """Get the entries for the right table of select-figure."""
+
+    pd.set_option('max_colwidth', 1000)
+
+    df = pd.DataFrame(columns=[  # Set the order of the columns
+        'CURATED-COFs ID', 'Article', 'Original Structure', 'Optimized Structure', 'CCS Workflow', 'Vers.'
+    ])
+
+    db_nodes_dict = get_db_nodes_dict()
+
+    for mat_id, mat_dict in db_nodes_dict.items():
+        new_row = {
+            'CURATED-COFs ID': mat_id,
+            'Article': doi_link(mat_dict),
+            'Original Structure': provenance_link(mat_dict['orig_cif']),
+            'Vers.': mat_dict['workflow_version'],
+        }
+
+        if 'opt_cif_ddec' in mat_dict:
+            new_row['Optimized Structure'] = provenance_link(mat_dict['opt_cif_ddec'])
+        else:
+            new_row['Optimized Structure'] = "N/A"
+
+        if 'appl_pecoal' in mat_dict:
+            new_row['CCS Workflow'] = detail_link(mat_id)
+        else:
+            new_row['CCS Workflow'] = "N/A"
+
+        df = df.append(new_row, ignore_index=True)
+
+    df = df.sort_values(by=['CURATED-COFs ID'])
+    df = df.reset_index(drop=True)
+    df.index += 1
+    return df
+
+
+# Get color palette
 myRdYlGn = (  # modified from Turbo256
     '#5dfb6f',
     '#61fc6c',
